+++ plugin/core/registry.py
@@ -1,40 +1,24 @@
-from .paths import simple_path
-from .promise import Promise
-from .protocol import CallHierarchyItem
-from .protocol import CallHierarchyPrepareParams
 from .protocol import Diagnostic
 from .protocol import Location
 from .protocol import LocationLink
 from .protocol import Point
-from .protocol import Request
-from .protocol import TextDocumentPositionParams
-from .protocol import TypeHierarchyItem
-from .protocol import TypeHierarchyPrepareParams
 from .sessions import AbstractViewListener
 from .sessions import Session
 from .tree_view import TreeDataProvider
-from .tree_view import TreeItem
 from .tree_view import TreeViewSheet
-from .typing import Callable, Optional, Any, Generator, Iterable, List, Union
-from .typing import cast
+from .typing import Optional, Any, Generator, Iterable, List, Union
 from .views import first_selection_region
 from .views import get_uri_and_position_from_location
-from .views import make_command_link
 from .views import MissingUriError
 from .views import point_to_offset
-from .views import SYMBOL_KINDS
-from .views import text_document_position_params
 from .views import uri_from_view
 from .windows import WindowManager
 from .windows import WindowRegistry
-from abc import ABCMeta
-from abc import abstractmethod
 from functools import partial
 import operator
 import sublime
 import sublime_api  # pyright: ignore[reportMissingImports]
 import sublime_plugin
-import weakref
 
 
 windows = WindowRegistry()
@@ -362,189 +346,3 @@ class LspCollapseTreeItemCommand(LspWindowCommand):
 
     def run(self, name: str, id: str) -> None:
         toggle_tree_item(self.window, name, id, False)
-
-
-HierarchyItem = Union[CallHierarchyItem, TypeHierarchyItem]
-
-
-class HierarchyDataProvider(TreeDataProvider):
-
-    def __init__(
-        self,
-        weaksession: 'weakref.ref[Session]',
-        request1: Callable[..., Request],
-        request2: Callable[..., Request],
-        request_handler1: Callable[..., List[HierarchyItem]],
-        request_handler2: Callable[..., List[HierarchyItem]],
-        direction: int,
-        root_elements: List[HierarchyItem]
-    ) -> None:
-        self.weaksession = weaksession
-        self.request1 = request1
-        self.request2 = request2
-        self.request_handler1 = request_handler1
-        self.request_handler2 = request_handler2
-        self.direction = direction
-        self.root_elements = root_elements
-        session = self.weaksession()
-        self.session_name = session.config.name if session else None
-
-    def get_children(self, element: Optional[HierarchyItem]) -> Promise[List[HierarchyItem]]:
-        if element is None:
-            return Promise.resolve(self.root_elements)
-        session = self.weaksession()
-        if not session:
-            return Promise.resolve([])
-        if self.direction == 1:
-            return session.send_request_task(self.request1({'item': element})).then(self.request_handler1)
-        if self.direction == 2:
-            return session.send_request_task(self.request2({'item': element})).then(self.request_handler2)
-        return Promise.resolve([])
-
-    def get_tree_item(self, element: HierarchyItem) -> TreeItem:
-        command_url = sublime.command_url('lsp_open_location', {
-            'location': {
-                'targetUri': element['uri'],
-                'targetRange': element['range'],
-                'targetSelectionRange': element['selectionRange']
-            },
-            'session_name': self.session_name,
-            'flags': sublime.ADD_TO_SELECTION | sublime.SEMI_TRANSIENT | sublime.CLEAR_TO_RIGHT
-        })
-        path = simple_path(self.weaksession(), element['uri'])
-        return TreeItem(
-            element['name'],
-            kind=SYMBOL_KINDS.get(element['kind'], sublime.KIND_AMBIGUOUS),
-            description=element.get('detail', ""),
-            tooltip="{}:{}".format(path, element['selectionRange']['start']['line'] + 1),
-            command_url=command_url
-        )
-
-
-def make_data_provider(
-    weaksession: 'weakref.ref[Session]', sheet_name: str, direction: int, response: List[HierarchyItem]
-) -> HierarchyDataProvider:
-    if sheet_name == "Call Hierarchy":
-        request1 = Request.incomingCalls
-        request2 = Request.outgoingCalls
-        handler1 = lambda response: [incoming_call['from'] for incoming_call in response] if isinstance(response, list) else []  # noqa: E501,E731
-        handler2 = lambda response: [outgoing_call['to'] for outgoing_call in response] if isinstance(response, list) else []  # noqa: E501,E731
-    elif sheet_name == "Type Hierarchy":
-        request1 = Request.supertypes
-        request2 = Request.subtypes
-        handler1 = handler2 = lambda response: response if isinstance(response, list) else []  # noqa: E731
-    else:
-        raise NotImplementedError('{} not implemented'.format(sheet_name))
-    return HierarchyDataProvider(weaksession, request1, request2, handler1, handler2, direction, response)
-
-
-def make_header(session_name: str, sheet_name: str, direction: int, root_elements: List[HierarchyItem]) -> str:
-    if sheet_name == "Call Hierarchy":
-        label = "Callers of…" if direction == 1 else "Calls from…"
-        tooltip = "Show outgoing calls" if direction == 1 else "Show Incoming Calls"
-    elif sheet_name == "Type Hierarchy":
-        label = "Supertypes of…" if direction == 1 else "Subtypes of…"
-        tooltip = "Show Subtypes" if direction == 1 else "Show Supertypes"
-    else:
-        raise NotImplementedError('{} not implemented'.format(sheet_name))
-    new_direction = 2 if direction == 1 else 1
-    return '{}: {} {}'.format(sheet_name, label, make_command_link('lsp_hierarchy_toggle', "⇄", {
-            'session_name': session_name,
-            'sheet_name': sheet_name,
-            'direction': new_direction,
-            'root_elements': root_elements
-        }, tooltip=tooltip))
-
-
-class LspHierarchyCommand(LspTextCommand, metaclass=ABCMeta):
-
-    @classmethod
-    @abstractmethod
-    def request(cls, params: TextDocumentPositionParams, view: sublime.View) -> Request:
-        """ A function that generates the initial request when this command is invoked. """
-        raise NotImplementedError()
-
-    def is_visible(self, event: Optional[dict] = None, point: Optional[int] = None) -> bool:
-        if self.applies_to_context_menu(event):
-            return self.is_enabled(event, point)
-        return True
-
-    def run(self, edit: sublime.Edit, event: Optional[dict] = None, point: Optional[int] = None) -> None:
-        self._window = self.view.window()
-        session = self.best_session(self.capability)
-        if not session:
-            return
-        position = get_position(self.view, event, point)
-        if position is None:
-            return
-        params = text_document_position_params(self.view, position)
-        session.send_request_async(
-            self.request(params, self.view), partial(self._handle_response_async, weakref.ref(session)))
-
-    def _handle_response_async(
-        self, weaksession: 'weakref.ref[Session]', response: Optional[List[HierarchyItem]]
-    ) -> None:
-        if not self._window or not self._window.is_valid():
-            return
-        if self.capability == 'callHierarchyProvider':
-            sheet_name = "Call Hierarchy"
-        elif self.capability == 'typeHierarchyProvider':
-            sheet_name = "Type Hierarchy"
-        else:
-            raise NotImplementedError('{} not implemented'.format(self.capability))
-        if not response:
-            self._window.status_message("{} not available".format(sheet_name))
-            return
-        session = weaksession()
-        if not session:
-            return
-        header = make_header(session.config.name, sheet_name, 1, response)
-        data_provider = make_data_provider(weaksession, sheet_name, 1, response)
-        new_tree_view_sheet(self._window, sheet_name, data_provider, header)
-        open_first(self._window, session.config.name, response)
-
-
-class LspHierarchyToggleCommand(LspWindowCommand):
-
-    def run(
-        self, session_name: str, sheet_name: str, direction: int, root_elements: List[HierarchyItem]
-    ) -> None:
-        session = self.session_by_name(session_name)
-        if not session:
-            return
-        header = make_header(session_name, sheet_name, direction, root_elements)
-        data_provider = make_data_provider(weakref.ref(session), sheet_name, direction, root_elements)
-        new_tree_view_sheet(self.window, sheet_name, data_provider, header)
-        open_first(self.window, session.config.name, root_elements)
-
-
-def open_first(window: sublime.Window, session_name: str, items: List[HierarchyItem]) -> None:
-    if items and window.is_valid():
-        item = items[0]
-        window.run_command('lsp_open_location', {
-            'location': {
-                'targetUri': item['uri'],
-                'targetRange': item['range'],
-                'targetSelectionRange': item['selectionRange']
-            },
-            'session_name': session_name,
-            'flags': sublime.ADD_TO_SELECTION | sublime.SEMI_TRANSIENT | sublime.CLEAR_TO_RIGHT
-        })
-
-
-class LspCallHierarchyCommand(LspHierarchyCommand):
-
-    capability = 'callHierarchyProvider'
-
-    @classmethod
-    def request(cls, params: TextDocumentPositionParams, view: sublime.View) -> Request:
-        return Request.prepareCallHierarchy(cast(CallHierarchyPrepareParams, params), view)
-
-
-class LspTypeHierarchyCommand(LspHierarchyCommand):
-
-    capability = 'typeHierarchyProvider'
-
-    @classmethod
-    def request(cls, params: TextDocumentPositionParams, view: sublime.View) -> Request:
-        return Request.prepareTypeHierarchy(cast(TypeHierarchyPrepareParams, params), view)
